fix(move golang package in directory and typo in readme):
