+++ binding/golang/main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"os"
-	"github.com/mohong122/ip2region/binding/golang"
+	"github.com/mohong122/ip2region/binding/golang/ip2region"
 	"bufio"
 	"fmt"
 	"strings"
